Feature #90: feat navigate storeDetail

+++ feature/home/src/main/java/com/hankki/feature/home/navigation/HomeNavigation.kt
@@ -16,7 +16,7 @@ fun NavController.navigateHome(navOptions: NavOptions) {
 fun NavGraphBuilder.homeNavGraph(
     paddingValues: PaddingValues,
     onShowSnackBar: (Int) -> Unit,
-    navigateStoreDetail: () -> Unit,
+    navigateStoreDetail: (Long) -> Unit,
     navigateToUniversitySelection: () -> Unit
 ) {
     composable<Home> {
